Fix int division error
When calculating bytes per frame, switch to casting to int at the end
rather than the middle, because bitDepth / 8 will usually be a fraction
and round to 0 and then bytesPerFrame is zero and that's a problem

+++ VideoHomeStorageTIB/VHSEncoder.cs
@@ -27,7 +27,7 @@ namespace VideoHomeStorage.TIB
         private static int numRows = vRows;
         private static int symbolWidth = streamWidth / numCols;
         private static int symbolHeight = streamHeight / numRows;
-        private static int bytesPerFrame = hBlocks * 8 * vRows * ((int)bitDepth / 8);
+        private static int bytesPerFrame = (int)(hBlocks * 8 * vRows * ((float)bitDepth / 8F));
 
         public static Bitmap Encode(byte[] data)
         {
